performance bug in yamlmodelrunner
marking agents to relocate with np.nan is a big performance issue.  I think this forces the type to float and then calling .loc with a float is a major performance issue.  Moving this back to marking relocating agents with -1 fixes the problem for now.  This might need a more permanent fix.

Also there was a warning in sqftproforma that needed to be addressed