Add ability to cache injectable results on inputs
If you have a utility function registered as an injectable but
it is not being automatically called you can now get caching of
results based on the inputs so long as the inputs are all hashable.
Use the memoize=True keyword along with autocall=False to get
this behavior.

Note that this only applies if you're using the utility via
the simulation framework (via injection or get_injectable).