change suffix on merge
This is something we probably should have done a long time ago - when you have columns defined on multiple tables in a merge - e.g. zone_id defined on parcels and buildings, then the result has zone_id_x and zone_id_y.  With this change you'll have zone_id and zone_id_y, so as long as the two zone_ids are the same after merging, this is probably what the user expects.